+++ apps/mobile/src/stores/preferences.ts
@@ -21,6 +21,7 @@ import {
 export const PREFERENCES_KEY = 'preferences-storage'
 
 export type PreferencesPayload = {
+  _experimental_video: boolean
   autoPlay: boolean
   blurNavigation: boolean
   blurNsfw: boolean
@@ -91,6 +92,7 @@ type State = PreferencesPayload & {
 export const usePreferences = create<State>()(
   persist(
     (set) => ({
+      _experimental_video: false,
       autoPlay: true,
       blurNavigation: true,
       blurNsfw: true,